[AIRFLOW-1314] seperated request classes
