+++ airflow/contrib/kubernetes/kubernetes_request_factory/__init__.py
@@ -1,0 +1,3 @@
+from .kubernetes_request_factory import *
+from .job_request_factory import *
+from .pod_request_factory import *
